+++ httpx/dispatch/wsgi.py
@@ -1,8 +1,8 @@
 import io
 import typing
 
+from .base import Dispatcher
 from ..config import CertTypes, TimeoutTypes, VerifyTypes
-from ..interfaces import Dispatcher
 from ..models import Request, Response
